+++ tests/audio.rs
@@ -91,6 +91,12 @@ fn test_monkeys_audio() {
 }
 
 #[test]
+fn test_mp3_url() {
+    let format = FileFormat::from_file("fixtures/audio/sample.m3u").unwrap();
+    assert_eq!(format, FileFormat::Mp3Url);
+}
+
+#[test]
 fn test_mpeg12_audio_layer3() {
     let format = FileFormat::from_file("fixtures/audio/sample.mp3").unwrap();
     assert_eq!(format, FileFormat::Mpeg12AudioLayer3);